Reddit/requirements.txt
praw
python-dotenv
SQLAlchemy